Upgrade to Upgrade to v2.13.0 (#2)
* Improved UI/UX (expert mode)
* updating unit tests
* add test case (show path and sign)

+++ app/src/common/tx.h
@@ -49,10 +49,10 @@ uint8_t *tx_get_buffer();
 const char *tx_parse();
 
 /// Return the number of items in the transaction
-tx_error_t tx_getNumItems(uint16_t *num_items);
+tx_error_t tx_getNumItems(uint8_t *num_items);
 
 /// Gets an specific item from the transaction (including paging)
-tx_error_t tx_getItem(int8_t displayIdx,
-                           char *outKey, uint16_t outKeyLen,
-                           char *outValue, uint16_t outValueLen,
-                           uint8_t pageIdx, uint8_t *pageCount);
+tx_error_t tx_getItem(uint8_t displayIdx,
+                      char *outKey, uint16_t outKeyLen,
+                      char *outValue, uint16_t outValueLen,
+                      uint8_t pageIdx, uint8_t *pageCount);
